Move SettingsFragment to a new file
