+++ AnkiDroid/src/main/java/com/ichi2/anki/preferences/NotificationsSettingsFragment.kt
@@ -21,7 +21,6 @@ import android.content.Intent
 import androidx.preference.ListPreference
 import androidx.preference.SwitchPreference
 import com.ichi2.anki.Preferences
-import com.ichi2.anki.Preferences.SettingsFragment
 import com.ichi2.anki.R
 import com.ichi2.anki.services.BootService.Companion.scheduleNotification
 import com.ichi2.anki.services.NotificationService
